refactor(modularization): Phase 3 Step 5b1 — collection_summary_service → KB domain + legacy shim (#16)
First of three sub-commits inside Step 5b per @符炫炜 msg=da460d8c
(Option P4 canonical: Protocol + DI). Step 5b1 moves the simplest of
the three KB services — ``collection_summary_service`` — because it
has no marketplace / search_pipeline cross-service dependencies. The
DI / Protocol ports work lands in 5b2 (collection_service) and 5b3
(document_service).

Changes

- ``git mv aperag/service/collection_summary_service.py`` →
  ``aperag/domains/knowledge_base/service/collection_summary_service.py``.
- New ``aperag/domains/knowledge_base/service/__init__.py`` placeholder.
- Inside the moved file: rewrite the single ``from aperag.db.models
  import ...`` block into two domain-local blocks, one per domain:
  - ``from aperag.domains.indexing.db.models import DocumentIndex,
    DocumentIndexStatus, DocumentIndexType``
  - ``from aperag.domains.knowledge_base.db.models import Collection,
    CollectionSummary, CollectionSummaryStatus, Document``
  Other imports (``aperag.config`` / ``aperag.db.ops`` /
  ``aperag.domains.indexing.summary_index`` /
  ``aperag.llm.completion.base_completion`` / ``aperag.schema.utils`` /
  ``aperag.tasks.reconciler``) are outside the G1 strict-ban target
  list and stay unchanged.
- New ``aperag/service/collection_summary_service.py`` legacy shim:
  ``from aperag.domains.knowledge_base.service.collection_summary_service
  import *`` plus explicit names for ``CollectionSummaryService`` and
  the module-level ``collection_summary_service`` instance so ``from
  aperag.service.collection_summary_service import
  collection_summary_service`` keeps working for every caller
  (``tasks/`` / ``domains/knowledge_graph/service.py`` /
  ``service/collection_service.py`` pre-5b2 / etc).

Gate status

- G1 strict-ban: ``aperag/domains/knowledge_base/service/`` has zero
  imports from ``aperag.db.models`` / ``aperag.service.*`` /
  ``aperag.schema.view_models``.
- G11 / G13 continue to hold (no class moves).
- ``uv run pytest tests/unit_test/`` → 669 passed, 29 skipped.
- ``make lint`` → all checks passed.

Out of scope — 5b2 / 5b3

- ``collection_service`` and ``document_service`` still live in
  ``aperag/service/`` and keep their marketplace / search_pipeline
  imports. The Protocol ports in
  ``aperag/domains/knowledge_base/ports.py`` + DI setters land when
  those services move (5b2 / 5b3).
- ``app.py`` bootstrap wire-up is deferred to 5b2/5b3 because no DI
  slots exist yet.

References: @符炫炜 msg=da460d8c (Option P4 canonical + 5b1/5b2/5b3
sub-split) + msg=4721535e (legacy service shim requirement).
